Replace uint8_t by int in MegaSet type property
Current SWIG version used complains about `uint8_t` type. This
changeset wraps that limitation in the public interface while keeping
an internal `uint8_t`

Additionally, use the opportunity to introduce an invalid value for
generic `MegaSet`s

Integration test updated accordingly

+++ src/megaapi.cpp
@@ -5619,7 +5619,7 @@ bool MegaApi::driveMonitorEnabled()
     return pImpl->driveMonitorEnabled();
 }
 
-void MegaApi::createSet(const char* name, uint8_t type, MegaRequestListener* listener)
+void MegaApi::createSet(const char* name, int type, MegaRequestListener* listener)
 {
     int options = CREATE_SET | (name ? OPTION_SET_NAME : 0);
     pImpl->putSet(INVALID_HANDLE, options, name, INVALID_HANDLE, type, listener);
